# artefacts/eap/a7/a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'LBAW Gameorama Web API'
  description: 'Web Resources Specification (A7) for Gameorama'
  contact:
    email: 'lbaw2133@gmail.com'
servers:
  - url: http://lbaw2133.lbaw-prod.fe.up.pt/
    description: Production server
...
paths:
...
  /api/member/{username}/posts/{page}:
    get:
      operationId: R116
      summary: "R116: Member's posts"
      description: "Shows member's posted news. Access: PUB"
      tags:
        - 'M01: Authentication and Individual Profile'
      parameters:
        - in: path
          name: username
          schema:
            type: string
          description: 'Specifies the username of the member'
          required: true
        - in: path
          name: page
          schema:
            type: string
          description: 'Specifies the number of the page for the query'
          required: true
      responses:
        '200':
          description: 'Ok. Shows [UI]'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NewsPost'
        '404':
          description: 'Member not found'
...
  /post/{id_post}:
    get:
      operationId: R207
      summary: 'R207: Shows a news post'
      description: 'Shows a post and its comments. Access: PUB'
      tags:
        - 'M02: Posts'
      parameters:
        - in: path
          name: id_post
          schema:
            type: integer
          description: 'Specifies the id of the post to be displayed'
          required: true
      responses:
        '200':
          description: 'Ok. Show [UI04](https://git.fe.up.pt/lbaw/lbaw2021/lbaw2133/-/wikis/er#ui04-post-page)'
        '400':
          description: 'Bad Request'
        '404':
          description: 'Invalid post id: Show [UI14](https://git.fe.up.pt/lbaw/lbaw2021/lbaw2133/-/wikis/er#ui14-error-404-page)'
    delete:
      operationId: R208
      summary: 'R208: Delete a post'
      description: 'Allows a member to delete a post. Access OWN, ADM'
      tags:
        - 'M02: Posts'
      parameters:
        - in: path
          name: id_post
          schema:
            type: integer
          description: 'Specifies the id of the post to be deleted'
          required: true
      responses:
        '302':
          description: 'Redirect after updating the news_post content.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful edition. Redirect to previous page.'
                302Error:
                  description: 'Failed edition. Redirect to previous page.'
...
  /api/post/{id_post}/vote:
    post:
      operationId: R213
      summary: 'R213: Rate a post'
      description: 'Allows a member to rate a news post. Access USR'
      tags:
        - 'M02: Posts'
      parameters:
        - in: path
          name: id_post
          schema:
            type: integer
          description: 'Specifies the id of the post to be rated'
          required: true
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Vote'
      responses:
        '200':
          description: 'Rated post successfully'
        '400':
          description: 'Bad Request'
        '403':
          description: 'Forbidden access'
        '404':
          description: 'Post not found'
        '302':
          description: 'Redirect visitor.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Redirect to login form.'
                  value: '/login'
    patch:
      operationId: R214
      summary: 'R214: Change vote on a news post'
      description: 'Allows a member to change their vote on a news post. Access: OWN'
      tags:
        - 'M02: Posts'
      parameters:
        - in: path
          name: id_post
          schema:
            type: integer
          description: 'Specifies the id of the post to be rated'
          required: true
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Vote'
      responses:
        '200':
          description: 'Vote edited successfully'
        '400':
          description: 'Bad Request'
        '403':
          description: 'Forbidden access'
        '404':
          description: 'Post not found'
    delete:
      operationId: R215
      summary: 'R215: Remove a vote on a news post'
      description: 'Allows a member to remove their vote on a post. Access: OWN'
      tags:
        - 'M02: Posts'
      parameters:
        - in: path
          name: id_post
          schema:
            type: integer
          description: 'Specifies the id of the post to remove the rating'
          required: true
      responses:
        '200':
          description: 'Vote deleted successfully'
        '400':
          description: 'Bad Request'
        '403':
          description: 'Forbidden access'
        '404':
          description: 'Post not found'
  /api/post/{id_post}/report:
    post:
      operationId: R216
      summary: 'R216: Report a post'
      description: 'Allows a member to report a news post. Access: USR'
      tags:
        - 'M02: Posts'
      parameters:
        - in: path
          name: id_post
          schema:
            type: integer
          description: 'Specifies the id of the post to be reported'
          required: true
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                body:
                  type: string
              required:
                - body
      responses:
        '200':
          description: 'Report successfully sent'
        '400':
          description: 'Bad Request'
        '403':
          description: 'Forbidden access'
        '404':
          description: 'Post not found'
...
  /api/search/posts:
    get:
      operationId: R603
      summary: 'R603: Search Posts API'
      description: 'Searches for posts and returns the result as JSON. Access: PUB'
      tags:
        - 'M06: Search'
      parameters:
        - in: query
          name: query
          schema:
            type: string
          required: true
          description: 'Specifies the query received'
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NewsPost'
        '400':
          description: 'Bad Request'
...
  /api/reports/posts:
    get:
      operationId: R702
      summary: 'R702: Reported Posts API'
      description: 'Returns reported posts as JSON. Access: ADM'
      tags:
        - 'M07: Administrator and Static Pages'
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NewsPost'
        '400':
          description: 'Bad Request'
        '403':
          description: 'Forbidden access'
...
components:
  schemas:
...
    Topic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
...
    NewsPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        date_time:
          type: string
          format: date-time
        aura:
          type: integer
        id_owner:
          type: integer
        username:
          type: string
        topics:
          type: array
          items:
            $ref: '#/components/schemas/Topic'
        images:
          type: array
          items:
            type: string
            format: byte
...
    Vote:
      type: object
      properties:
        vote:
          type: boolean
          enum: [true,false]
...